Add files via upload
The project is based on a road trip planner, when a user enters the start date, end date and the name of the place, it will show the path to be taken with the total miles calculated.
The nearby fuel stations are also displayed within a 10km radius. 
Also displays the hotels that can be booked for the dates entered. 

Technologies used are:
AJAX patterns : multistage download, submission throttling 
CSS, HTML, Javascript & PHP for building the website.